+++ frontend/app/src/components_mobile/App.svelte
@@ -13,27 +13,17 @@
     import { incomingVideoCall } from "@stores/video";
     import { broadcastLoggedInUser } from "@stores/xframe";
     import "@utils/markdown";
-    import { getProxyAdjustedBlobUrl } from "@utils/media";
     import {
         expectNewFcmToken,
         expectNotificationTap,
         expectPushNotifications,
-        expectShareTarget,
         expectWindowInsetChange,
-        type ShareTarget,
     } from "@utils/native/notification_channels";
-    import type { Share } from "@utils/share";
+    import { expectShareTarget, handleShareTarget } from "@utils/native/share_target";
     import "@utils/scream";
     import { portalState } from "component-lib";
     import {
-        allChatsStore,
-        allUsersStore,
         type ChatIdentifier,
-        type ChatSummary,
-        communitiesStore,
-        compareChats,
-        LazyFile,
-        localUpdates,
         OpenChat,
         type VideoCallType,
         botState,
@@ -41,18 +31,16 @@
         fontSize,
         identityStateStore,
         inititaliseLogger,
-        publish,
         routeForChatIdentifier,
         routeForScope,
         subscribe,
     } from "openchat-client";
-    import { convertFileSrc } from "@tauri-apps/api/core";
-    import { derived } from "svelte/store";
     import page from "page";
     import { onMount, setContext } from "svelte";
     import { overrideItemIdKeyNameBeforeInitialisingDndZones } from "svelte-dnd-action";
     import { _, isLoading } from "svelte-i18n";
-    import { getFcmToken, svelteReady, updateChatShortcuts } from "tauri-plugin-oc-api";
+    import { getFcmToken, svelteReady } from "tauri-plugin-oc-api";
+    import { startChatShortcutPusher } from "@stores/chatShortcuts";
     import Head from "./Head.svelte";
     import NotificationsBar from "./home/NotificationsBar.svelte";
     import ActiveCall from "./home/video/ActiveCall.svelte";
@@ -191,116 +179,6 @@
         setStatusAndNavBarSizesForNativeApp();
     }
 
-    // Shortcut ids round-trip through the Android shortcut system as opaque
-    // strings, so we need a kind-prefixed encoding to reverse them back into
-    // a ChatIdentifier when the share arrives. chatIdentifierToString in
-    // openchat-shared drops the kind, which makes direct vs group ids
-    // indistinguishable on the way back.
-    function chatIdToShortcutId(id: ChatIdentifier): string {
-        switch (id.kind) {
-            case "direct_chat":
-                return `d:${id.userId}`;
-            case "group_chat":
-                return `g:${id.groupId}`;
-            case "channel":
-                return `c:${id.communityId}:${id.channelId}`;
-        }
-    }
-
-    function shortcutIdToChatId(s: string): ChatIdentifier | undefined {
-        const colon = s.indexOf(":");
-        if (colon < 0) return undefined;
-        const kind = s.slice(0, colon);
-        const rest = s.slice(colon + 1);
-        switch (kind) {
-            case "d":
-                return { kind: "direct_chat", userId: rest };
-            case "g":
-                return { kind: "group_chat", groupId: rest };
-            case "c": {
-                const sep = rest.lastIndexOf(":");
-                if (sep < 0) return undefined;
-                const channelId = Number(rest.slice(sep + 1));
-                if (!Number.isFinite(channelId)) return undefined;
-                return {
-                    kind: "channel",
-                    communityId: rest.slice(0, sep),
-                    channelId,
-                };
-            }
-            default:
-                return undefined;
-        }
-    }
-
-    function shareToChat(chatId: ChatIdentifier, shareTarget: ShareTarget) {
-        // Set the draft synchronously — the chat doesn't have to exist in
-        // chatSummariesStore yet; the draft is keyed by id and the composer
-        // picks it up reactively once the chat is selected.
-        const text = shareTarget.text ?? "";
-        if (text.length > 0) {
-            localUpdates.draftMessages.setTextContent({ chatId }, text);
-        }
-        const firstFile = shareTarget.files[0];
-        if (firstFile) {
-            const lazy = LazyFile.fromUrl(
-                convertFileSrc(firstFile.path),
-                firstFile.name,
-                firstFile.mimeType ?? "application/octet-stream",
-                firstFile.size,
-            );
-            client
-                .messageContentFromFile(lazy as unknown as File)
-                .then((content) =>
-                    localUpdates.draftMessages.setAttachment({ chatId }, content),
-                )
-                .catch((err) => console.error("Failed to attach shared file", err));
-        }
-        // Defer the navigation: on cold start, the share-target event can
-        // arrive while Home.svelte is still doing its initial route resolution
-        // (which pageReplaces the home_route to the default scope). Pushing
-        // through a macrotask lets that settle first, so our chat route wins.
-        setTimeout(() => page(routeForChatIdentifier($chatListScopeStore.kind, chatId)));
-    }
-
-    function handleShareTarget(shareTarget: ShareTarget) {
-        // Direct Share fast-path: the share came from a chat shortcut we
-        // pushed via updateChatShortcuts, so we know the destination and
-        // can skip the picker entirely.
-        if (shareTarget.shortcutId) {
-            const chatId = shortcutIdToChatId(shareTarget.shortcutId);
-            if (chatId !== undefined) {
-                shareToChat(chatId, shareTarget);
-                return;
-            }
-            // Unrecognised shortcut id — fall through to the picker.
-        }
-
-        // Generic share-sheet path: reuse the existing in-app "share message"
-        // flow. SlidingModals subscribes to "shareMessage" and renders
-        // ShareMessageModal, which wraps SelectChatModal and pre-fills the
-        // chosen chat's draft via localUpdates.draftMessages.
-        const text = shareTarget.text ?? "";
-        // Wrap each shared file path in a LazyFile so the existing
-        // messageContentFromFile path can stream bytes through Tauri's asset
-        // protocol instead of buffering them across the IPC boundary.
-        const files = shareTarget.files.map((f) =>
-            LazyFile.fromUrl(
-                convertFileSrc(f.path),
-                f.name,
-                f.mimeType ?? "application/octet-stream",
-                f.size,
-            ),
-        );
-        const share: Share = {
-            title: undefined,
-            text: text.length > 0 ? text : undefined,
-            url: undefined,
-            files: files as unknown as File[],
-        };
-        publish("shareMessage", share);
-    }
-
     // Sets up push notifications and FCM token management for native apps
     function setupNativeApp() {
         const addFcmToken = (token: string) => {
@@ -321,7 +199,7 @@
             // Listen for content shared into OpenChat from the system share sheet
             // (or via a Direct Share chat shortcut). Cold-start shares are queued
             // by the native side and delivered once svelteReady fires below.
-            expectShareTarget(handleShareTarget).catch(console.error);
+            expectShareTarget((share) => handleShareTarget(client, share)).catch(console.error);
 
             // Expect FCM token refreshes
             expectNewFcmToken(addFcmToken);
@@ -346,82 +224,8 @@
 
             // Push the top recent chats to the Android Sharing Shortcuts API
             // so they appear directly in the system share sheet (like
-            // WhatsApp's per-chat tiles). chatSummariesListStore re-fires on
-            // every message in any chat, but the top-N identity changes far
-            // less often — we dedupe by a stringified key to avoid hammering
-            // the native side (each push downloads avatars via Coil).
-            //
-            // Cap matches the typical Android Direct Share row (4 tiles).
-            const SHORTCUT_COUNT = 4;
-            const topShortcutsStore = derived(
-                [allChatsStore, allUsersStore, communitiesStore],
-                ([chats, users, communities]) => {
-                    // allChatsStore is a ChatMap across every scope (direct
-                    // chats, groups, channels in any community) so users see
-                    // the right tiles regardless of which scope the app is
-                    // currently in. compareChats matches the comparator the
-                    // in-app chat list uses internally.
-                    //
-                    // Filter out chats we can't actually send a message to
-                    // (OpenChatBot, proposal bot, read-only groups, etc.)
-                    // BEFORE slicing — otherwise a non-sendable chat near the
-                    // top would waste a tile slot.
-                    const sorted: ChatSummary[] = [...chats.values()].sort(compareChats);
-                    return sorted
-                        .filter((chat) => client.canSendMessage(chat.id, "message"))
-                        .slice(0, SHORTCUT_COUNT)
-                        .map((chat) => {
-                            if (chat.kind === "direct_chat") {
-                                const them = users.get(chat.them.userId);
-                                return {
-                                    id: chatIdToShortcutId(chat.id),
-                                    name: client.displayName(them),
-                                    // Adjust dev-mode localhost canister URLs so the
-                                    // native side can fetch the avatar bytes via Coil.
-                                    avatarUrl: getProxyAdjustedBlobUrl(client.userAvatarUrl(them)),
-                                };
-                            }
-                            if (chat.kind === "channel") {
-                                // Prefix channel tiles with the community name
-                                // so the user can tell which community the
-                                // channel belongs to (the channel name alone
-                                // is often ambiguous across communities).
-                                const community = communities.get({
-                                    kind: "community",
-                                    communityId: chat.id.communityId,
-                                });
-                                const name = community
-                                    ? `${community.name}#${chat.name}`
-                                    : chat.name;
-                                return {
-                                    id: chatIdToShortcutId(chat.id),
-                                    name,
-                                    avatarUrl: getProxyAdjustedBlobUrl(
-                                        client.groupAvatarUrl(chat),
-                                    ),
-                                };
-                            }
-                            return {
-                                id: chatIdToShortcutId(chat.id),
-                                name: chat.name,
-                                avatarUrl: getProxyAdjustedBlobUrl(client.groupAvatarUrl(chat)),
-                            };
-                        });
-                },
-            );
-            let lastPushedShortcutsKey = "";
-            topShortcutsStore.subscribe((chats) => {
-                // Skip the transient empty emit during chat-list hydration —
-                // pushing an empty list would prune any existing shortcuts,
-                // leaving a window where the system share sheet has none.
-                if (chats.length === 0) return;
-                const key = JSON.stringify(chats);
-                if (key === lastPushedShortcutsKey) return;
-                lastPushedShortcutsKey = key;
-                updateChatShortcuts({ chats }).catch((err) =>
-                    console.error("Failed to update chat shortcuts", err),
-                );
-            });
+            // WhatsApp's per-chat tiles).
+            startChatShortcutPusher(client);
 
             // Inform the native android app that svelte code is ready! SetTimeout
             // delays the fn execution until the call stack is empty, just to
